requirements.txt
azure-functions
azure-functions-worker
grpcio==1.14.1
grpcio-tools==1.14.1
protobuf==3.6.1
six==1.11.0
pandas==0.24.2
numpy==1.16.2
requests==2.21.0
nltk==3.4